[AIRFLOW-5885] List of tests is generated by pre-commit not Docker build
The list of tests does not change that often, yet it could not be generated
in Docker before all files were added (due to imports etc.). Therefore it is much better
solution to generate the list of tests in pre-commit - only when any of the test .py files
change and commit it together with the sources.

Also - since we are modifying Dockerfile, it turned out that "airflow" user is not needed
in the CI image any more so I removed it to save additional Dockerfile modification.