Remove "proposed remediation" from the template
This field adds no value, if there is a proposed remediation at
the time of issue creation (rarely) then it should be part
of the description and the remediation proposals should follow
in the discussion.

I created ~20 issues while moving them from security@ and except
one case I never filled that one (not felt the need) - and in the
one that I did use it #2 initially, it could have been easily
part of the description.