Adapt the issue managment process to status quo
I have a few updates to the process description to adapt to what has been established as the "status-quo" way of our process handling. 

Apparently when you prepare a new "housing area", the best way to build the paths is to leave grass and wait a bit for the people living there,  to just walk as they need and then apparently the most useful walking paths emerge automatically by the grass disappearing where people are actually walking - and only then make those paths "official". 

This is an attempt to "capture" the waliking paths observed as "officially documented ones".

* when issue is obvious we should directly respond and explain without extra overhead of creating issue/discussing

* we started to use "Not yet announced" label for issues that are already fixed but not yet announced, I think it's a great way for everyone to keep track the status rather than closing the issue and mark the milestone.

* I think it's a good idea to respond to the reporter (like @pierrejeambrun did for the last announcement) that we are looking at the report at the moment we create an issue. This will give the reporter confirmation that the report has been received and let them know the issue is being taken care of - which will save extra questions and keep the impatient reporters from nagging us.

Looing forward to reviews.