Temporarily switch to public github runners
We are having problems with self-hosted runners and until we fix
the problems we switch main builds to only run on main with
regular schedule 1x day to serve as a canary build. They will run
longer than on self-hosted runners.

We also add option to trigger the workflow manually from GitHub
interface (by maintainers).