Update the canned response for connection-induced RCE and similar
The text was more about "test connection" feature than the connection
so it's better to make it a bit more generic, to respond to the users
who report any issues related with connection extras.

Also added an information that the reporter is free to report a
regular issue - or even better - make a PR to change the behaviour.