Improve documentation on how maintainers can debug/modify CI behaviour
The description grew from small number of labels to a "wall of text"
that was difficult to read. This PR reformats it in the form of
table that makes it far easier to see actions that the maintainer
can do and what labels should be set for each of them.

Also for removing dependencies, you need to set both:

* `disable docker cache` label
* incerase DEPENDENCIES_EPOCH_NUMBER in `Dockerfile.ci`

Comments and documentation in both places has been updated to
reflect it.